+++ src/NetStandard/E.Standard.WebMapping.Core/KnownSRef.cs
@@ -1,0 +1,9 @@
+﻿namespace E.Standard.WebMapping.Core;
+
+static public class KnownSRef
+{
+    public const int WGS84 = 4326;
+    public const int WebMercator = 3857;
+
+    public static readonly int[] WebMercatorIds = new int[] { WebMercator };
+}